General cleanup (upgrade to quasar-0.6.2)
