+++ src/main/java/com/github/vy/QuasarInstrumentorMojo.java
@@ -1,5 +1,10 @@
 package com.github.vy;
 
+import co.paralleluniverse.fibers.instrument.DefaultSuspendableClassifier;
+import co.paralleluniverse.fibers.instrument.Log;
+import co.paralleluniverse.fibers.instrument.LogLevel;
+import co.paralleluniverse.fibers.instrument.MethodDatabase;
+import co.paralleluniverse.fibers.instrument.QuasarInstrumentor;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
@@ -7,6 +12,12 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.Queue;
@@ -16,7 +27,6 @@ import java.util.Queue;
  */
 @Mojo(name = "quasar", defaultPhase = LifecyclePhase.PROCESS_CLASSES)
 public class QuasarInstrumentorMojo extends AbstractMojo {
-
     /**
      * Build directory path.
      */
@@ -30,7 +40,7 @@ public class QuasarInstrumentorMojo extends AbstractMojo {
     protected boolean check = false;
 
     /**
-     * Enable to display the name of each processed class and all suspendable method calles.
+     * Enable to display the name of each processed class and all suspendable method calls.
      */
     @Parameter
     protected boolean verbose = false;
@@ -61,23 +71,97 @@ public class QuasarInstrumentorMojo extends AbstractMojo {
             throw new MojoExecutionException("Invalid build directory: " + buildDirectory);
 
         // Create a Quasar instrumentor.
-        QuasarWrapper quasar = new QuasarWrapper(
-                this, check, verbose, debug, allowMonitors, allowBlocking);
+        final QuasarInstrumentor instrumentor;
+        try {
+            final ClassLoader cl = new URLClassLoader(new URL[]{buildDirectory.toURI().toURL()}, getClass().getClassLoader());
+            instrumentor = new QuasarInstrumentor(true, cl, new DefaultSuspendableClassifier(cl));
+        } catch (MalformedURLException e) {
+            throw new AssertionError(e);
+        }
+
+        instrumentor.setCheck(check);
+        instrumentor.setVerbose(verbose);
+        instrumentor.setDebug(debug);
+        instrumentor.setAllowBlocking(allowBlocking);
+        instrumentor.setAllowMonitors(allowMonitors);
+        instrumentor.setLog(new Log() {
+
+            @Override
+            public void log(LogLevel level, String msg, Object... args) {
+                final String message = String.format(msg, args);
+                switch (level) {
+                    case DEBUG:
+                        getLog().debug(message);
+                        break;
+                    case INFO:
+                        getLog().info(message);
+                        break;
+                    case WARNING:
+                        getLog().warn(message);
+                        break;
+                    default:
+                        getLog().error(message);
+                        break;
+                }
+            }
+
+            @Override
+            public void error(String s, Exception e) {
+                getLog().error(s, e);
+            }
+
+        });
 
         // Traverse the build directory and feed the class files to the instrumentor.
-        Queue<File> files = new LinkedList<>(Collections.singleton(buildDirectory));
+        final Queue<File> files = new LinkedList<>(Collections.singleton(buildDirectory));
         while (!files.isEmpty()) {
             File file = files.poll();
             if (file.isDirectory()) {
                 File[] dirFiles = file.listFiles();
                 if (dirFiles != null)
                     Collections.addAll(files, dirFiles);
-            }
-            else if (file.getName().endsWith(".class")) quasar.checkClass(file);
+            } else if (file.getName().endsWith(".class"))
+                instrumentor.checkClass(file);
         }
 
-        // Let the instrumentor kick in.
-        quasar.instrumentClasses();
+        instrumentClasses(instrumentor);
     }
 
+    protected void logDebug(String fmt, Object... args) {
+        getLog().debug(String.format(fmt, args));
+    }
+
+    protected void logInfo(String fmt, Object... args) {
+        getLog().info(String.format(fmt, args));
+    }
+
+    protected void logWarn(String fmt, Object... args) {
+        getLog().warn(String.format(fmt, args));
+    }
+
+    protected void logError(String s, Exception e) {
+        getLog().error(s, e);
+    }
+
+    public void instrumentClasses(QuasarInstrumentor instrumentor) throws MojoExecutionException {
+        logInfo("Instrumenting %d classes...", instrumentor.getWorkList().size());
+        for (MethodDatabase.WorkListEntry wle : instrumentor.getWorkList())
+            instrumentClass(instrumentor, wle);
+    }
+
+    private void instrumentClass(QuasarInstrumentor instrumentor, MethodDatabase.WorkListEntry entry) throws MojoExecutionException {
+        if (!instrumentor.shouldInstrument(entry.name))
+            return;
+        try {
+            try (FileInputStream fis = new FileInputStream(entry.file)) {
+                String className = entry.name.replace('.', '/');
+                byte[] newClass = instrumentor.instrumentClass(className, fis);
+                try (FileOutputStream fos = new FileOutputStream(entry.file)) {
+                    fos.write(newClass);
+                }
+            }
+        } catch (IOException ex) {
+            throw new MojoExecutionException("Instrumenting file " + entry.file, ex);
+        }
+    }
 }
